requirements.txt
coverage==7.3.2
django==2.2.6
pillow==7.0.0
sorl-thumbnail==12.6.3
Unidecode==1.1.1 
djangorestframework==3.12.2
WeasyPrint==52.1
python-dotenv==0.15.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
